requirements.txt
requests==2.31.0
pydantic==2.5.2
pydantic-settings==2.1.0
streamlit==1.29.0
pandas==2.1.4
plotly-express==0.4.1